Consolidate start port forward route to use injection token for registering route
